Release 0.2.0
From now on this crate won't build libcbqn shared object by default. It
is discouraged by the Cargo book[1].

Now that the direct dependency to CBQN is cut, I changed my mind about
the versioning of this crate. It does not necessarily follow CBQN's
versioning, but will somewhat resemble that. I doubt there will be much
changes in the bqnffi.h that this crate creates bindings to, so updating
this from now on isn't going to happen very often anyway.

[1]: https://doc.rust-lang.org/cargo/reference/build-scripts.html#rustc-link-search